ml/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "concert-scraper-ml"
version = "0.1.0"
description = "Transformer-only metadata extraction pipeline for concert-scraper"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0.1",
  "numpy>=1.26.4",
  "datasets>=2.20.0",
  "transformers>=4.44.2",
  "torch>=2.4.1",
  "accelerate>=0.34.2",
  "sentencepiece>=0.2.0",
  "scikit-learn>=1.5.1",
  "pandas>=2.2.2",
  "tqdm>=4.66.5"
]

[tool.setuptools]
py-modules = []
